ref(project-creation): Fold integration selector into ScmIntegrationConnect
Consolidate ScmIntegrationConnectMultiOrg back into ScmIntegrationConnect
behind an allowIntegrationSwitching prop. The two components were ~90%
identical (data fetching, loading/error, install, pills, repo selector); the
only divergence was the connected-state header.

When allowIntegrationSwitching is set, the connected state renders the
integration selector dropdown; otherwise it renders the static "Connected
to ..." text. Onboarding keeps the default (static text) and is unchanged;
project creation opts in.

Refs VDY-25
