--- static/app/views/onboarding/components/scmIntegrationConnectMultiOrg.tsx
@@ -1,160 +1,0 @@
-import {useCallback, useEffect} from 'react';
-import {motion} from 'framer-motion';
-
-import {Button} from '@sentry/scraps/button';
-import {Flex, Stack, type StackProps} from '@sentry/scraps/layout';
-import {Text} from '@sentry/scraps/text';
-
-import {LoadingIndicator} from 'sentry/components/loadingIndicator';
-import {t} from 'sentry/locale';
-import type {Integration, Repository} from 'sentry/types/integrations';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import {useOrganization} from 'sentry/utils/useOrganization';
-import {SCM_STEP_CONTENT_WIDTH} from 'sentry/views/onboarding/consts';
-
-import type {ScmAnalyticsFlow} from './scmAnalyticsFlow';
-import {ScmIntegrationSelect} from './scmIntegrationSelect';
-import {ScmProviderPills} from './scmProviderPills';
-import {ScmRepoSelector} from './scmRepoSelector';
-import {useMultiPlatformDetectionTest} from './useMultiPlatformDetectionTest';
-import {useScmPlatformDetection} from './useScmPlatformDetection';
-import {useScmProviders} from './useScmProviders';
-
-const STEP_VIEWED_EVENT = {
-  onboarding: 'onboarding.scm_connect_step_viewed',
-  'project-creation': 'project_creation.scm_connect_step_viewed',
-} as const;
-
-const INTEGRATION_SELECTED_EVENT = {
-  onboarding: 'onboarding.scm_connect_integration_selected',
-  'project-creation': 'project_creation.scm_connect_integration_selected',
-} as const;
-
-interface ScmIntegrationConnectMultiOrgProps {
-  analyticsFlow: ScmAnalyticsFlow;
-  // Fired once per user-driven repo change so callers can invalidate state
-  // derived from the repo (platform, features, created project). See
-  // ScmRepoSelector for why this is separate from onRepositoryChange.
-  onClearDerivedState: () => void;
-  onIntegrationChange: (integration: Integration | undefined) => void;
-  onRepositoryChange: (repo: Repository | undefined) => void;
-  selectedIntegration: Integration | undefined;
-  selectedRepository: Repository | undefined;
-  maxWidth?: StackProps['maxWidth'];
-}
-
-/**
- * Variant of `ScmIntegrationConnect` that lets the user pick which connected
- * integration (provider + org/account) to search repositories within, instead
- * of defaulting to a single one. Renders the provider install pills when no
- * integration is connected, or the integration selector + repo selector when
- * one or more are. Owns integration data fetching, platform detection
- * pre-warming, and the `scm_connect_step_viewed` analytic.
- *
- * Designed for the SCM-first project creation surface (`ScmCreateProject`).
- * Like `ScmIntegrationConnect`, it does NOT render surrounding chrome (heading,
- * footer); hosts compose that around it.
- */
-export function ScmIntegrationConnectMultiOrg({
-  analyticsFlow,
-  onClearDerivedState,
-  onIntegrationChange,
-  onRepositoryChange,
-  selectedIntegration,
-  selectedRepository,
-  maxWidth = SCM_STEP_CONTENT_WIDTH,
-}: ScmIntegrationConnectMultiOrgProps) {
-  const organization = useOrganization();
-  const {
-    scmProviders,
-    activeIntegrations,
-    isPending,
-    isError,
-    refetch,
-    refetchIntegrations,
-  } = useScmProviders();
-
-  // Pre-warm platform detection so results are cached when the user advances
-  useScmPlatformDetection(selectedRepository);
-
-  // Measurement call, to judge latency (no UI impact).
-  useMultiPlatformDetectionTest(selectedRepository);
-
-  // Derive integration from explicit selection, falling back to the first
-  // active integration so the repo selector has something to search.
-  const effectiveIntegration = selectedIntegration ?? activeIntegrations[0];
-
-  useEffect(() => {
-    trackAnalytics(STEP_VIEWED_EVENT[analyticsFlow], {organization});
-  }, [organization, analyticsFlow]);
-
-  const handleInstall = useCallback(
-    (data: Integration) => {
-      onIntegrationChange(data);
-      onRepositoryChange(undefined);
-      refetchIntegrations();
-    },
-    [onIntegrationChange, onRepositoryChange, refetchIntegrations]
-  );
-
-  // Switching integrations invalidates the selected repo (repos are scoped to
-  // an integration) and everything derived from it (platform, features, form).
-  const handleIntegrationSelect = useCallback(
-    (integration: Integration) => {
-      onClearDerivedState();
-      onIntegrationChange(integration);
-      onRepositoryChange(undefined);
-      trackAnalytics(INTEGRATION_SELECTED_EVENT[analyticsFlow], {
-        organization,
-        provider: integration.provider.key,
-      });
-    },
-    [
-      analyticsFlow,
-      organization,
-      onClearDerivedState,
-      onIntegrationChange,
-      onRepositoryChange,
-    ]
-  );
-
-  if (isPending) {
-    return (
-      <Flex justify="center" align="center">
-        <LoadingIndicator mini />
-      </Flex>
-    );
-  }
-
-  if (isError) {
-    return (
-      <Stack gap="lg" align="center">
-        <Text variant="muted">{t('Failed to load integrations.')}</Text>
-        <Button onClick={() => refetch()}>{t('Retry')}</Button>
-      </Stack>
-    );
-  }
-
-  return effectiveIntegration ? (
-    <MotionStack key="with-integration" gap="md" width="100%" maxWidth={maxWidth}>
-      <ScmIntegrationSelect
-        integrations={activeIntegrations}
-        selectedIntegration={effectiveIntegration}
-        onChange={handleIntegrationSelect}
-      />
-      <ScmRepoSelector
-        analyticsFlow={analyticsFlow}
-        integration={effectiveIntegration}
-        selectedRepository={selectedRepository}
-        onRepositoryChange={onRepositoryChange}
-        onClearDerivedState={onClearDerivedState}
-      />
-    </MotionStack>
-  ) : (
-    <MotionStack key="without-integration" gap="2xl" width="100%" maxWidth={maxWidth}>
-      <ScmProviderPills providers={scmProviders} onInstall={handleInstall} />
-    </MotionStack>
-  );
-}
-
-const MotionStack = motion.create(Stack);
